% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/High_School_Completion_Rate_DATA.R
\docType{data}
\name{High_School_Completion_Rate}
\alias{High_School_Completion_Rate}
\title{High School Completion Rate in Baltimore City}
\format{
## `High_School_Completion_Rate`
A data frame with 55 obs. of  11 variables:
\describe{
  \item{CSA2010 }{Area Name in 2010 (Character String)}
  \item{compl10}{The percentage of 12th graders in a school year that successfully completed high school in 2010 (Numeric)}
  \item{compl11}{The percentage of 12th graders in a school year that successfully completed high school in 2011 (Numeric)}
  \item{compl12}{The percentage of 12th graders in a school year that successfully completed high school in 2012 (Numeric)}
  \item{compl13}{The percentage of 12th graders in a school year that successfully completed high school in 2013 (Numeric)}
  \item{compl14}{The percentage of 12th graders in a school year that successfully completed high school in 2014 (Numeric)}
  \item{compl15}{The percentage of 12th graders in a school year that successfully completed high school in 2015 (Numeric)}
  \item{compl16}{The percentage of 12th graders in a school year that successfully completed high school in 2016 (Numeric)}
  \item{compl17}{The percentage of 12th graders in a school year that successfully completed high school in 2017 (Numeric)}
  \item{compl18}{The percentage of 12th graders in a school year that successfully completed high school in 2018 (Numeric)}
  \item{compl19}{The percentage of 12th graders in a school year that successfully completed high school in 2019 (Numeric)}
  \item{compl20}{The percentage of 12th graders in a school year that successfully completed high school in 2020 (Numeric)}
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::high-school-completion-rate-community-statistical-area/about>
}
\usage{
High_School_Completion_Rate
}
\description{
The percentage of 12th graders in a school year that successfully completed high school out of all 12th graders within an area.
Completers are identified as completing their program of study at the high school level and satisfying the graduation requirements for a Maryland High School Diploma or the requirements for a Maryland Certificate of Program Completion.
}
\keyword{datasets}
