% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infant_Mortality_Rate_DATA.R
\docType{data}
\name{Infant_Mortality_Rate}
\alias{Infant_Mortality_Rate}
\title{Infant Mortality Rate in Baltimore City}
\format{
## `Infant_Mortality_Rate`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::infant-mortality-rate/about>
}
\usage{
Infant_Mortality_Rate
}
\description{
The number of infant deaths (babies under one year of age) per 1,000 live births within the area in a five year period.
This is the most stable and commonly measured indicator of mortality in this age group.
}
\keyword{datasets}
