% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kindergarten_Readiness_DATA.R
\docType{data}
\name{Kindergarten_Readiness}
\alias{Kindergarten_Readiness}
\title{Kindergarten Readiness Assessment (KRA) results in Baltimore City}
\format{
## `Kindergarten_Readiness`
A data frame with 55 obs. of  3 variables:
\describe{
  \item{CSA2010 }{Area Name in 2010 (Character String)}
  \item{kra16}{The Kindergarten Readiness Assessment (KRA) measure in 2016 (Numeric)}
  \item{kra17}{The Kindergarten Readiness Assessment (KRA) measure in 2017 (Numeric)}
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::kindergarten-readiness/about>
}
\usage{
Kindergarten_Readiness
}
\description{
The Kindergarten Readiness Assessment (KRA) measure children's readiness to do kindergarten work.
The KRA is a developmentally appropriate assessment tool that measures of incoming public-school kindergarteners across four learning domains.
A child is considered to be ready for kindergarten if assessed as "demonstrating" the foundational skills and behaviors that prepare him/her for curriculum based on the kindergarten standards.
}
\keyword{datasets}
