% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Life_Expectancy_DATA.R
\docType{data}
\name{Life_Expectancy}
\alias{Life_Expectancy}
\title{Life Expectancy in Baltimore City}
\format{
## `Life_Expectancy`
A data frame with 55 obs. of  9 variables:
\describe{
  \item{CSA2010 }{Area Name (Character String)}
  \item{lifexp11}{The average number of years a newborn can expect to live in 2011  (Numeric)}
  \item{lifexp12}{The average number of years a newborn can expect to live in 2012  (Numeric)}
  \item{lifexp13}{The average number of years a newborn can expect to live in 2013  (Numeric)}
  \item{lifexp14}{The average number of years a newborn can expect to live in 2014  (Numeric)}
  \item{lifexp15}{The average number of years a newborn can expect to live in 2015  (Numeric)}
  \item{lifexp16}{The average number of years a newborn can expect to live in 2016  (Numeric)}
  \item{lifexp17}{The average number of years a newborn can expect to live in 2017  (Numeric)}
  \item{lifexp18}{The average number of years a newborn can expect to live in 2018  (Numeric)}
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::life-expectancy-community-statistical-area/about>
}
\usage{
Life_Expectancy
}
\description{
The average number of years a newborn can expect to live, assuming he or she experiences the currently prevailing rates of death through their lifespan.
}
\keyword{datasets}
