% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open_Checkbook_FY2020_Dataset_DATA.R
\docType{data}
\name{Open_Checkbook_FY2020_Dataset}
\alias{Open_Checkbook_FY2020_Dataset}
\title{Baltimore City Open Checkbook FY2020 Dataset}
\format{
## `Open_Checkbook_FY2020_Dataset`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://data.baltimorecity.gov/datasets/baltimore::baltimore-city-employee-salaries-2/about>
}
\usage{
Open_Checkbook_FY2020_Dataset
}
\description{
This dataset contains city expenses with vendors/contractors. The first release includes only fiscal year 2020. Fields include date of transaction, name of vendor, categorizations of spending, agency, and the amount of the expense.
}
\keyword{datasets}
