% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perc_Pop_School_Employed_DATA.R
\docType{data}
\name{Perc_Pop_School_Employed}
\alias{Perc_Pop_School_Employed}
\title{Percentage of Population aged 16-19 in School or Employed in Baltimore City}
\format{
## `Perc_Pop_School_Employed`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percentage-of-population-aged-16-19-in-school-or-employed-community-statistical-area/about>
}
\usage{
Perc_Pop_School_Employed
}
\description{
The percentage of persons aged 16 to 19 who are in school and/or are employed out of all persons in their age cohort
broken down by community statistical area regions.
}
\keyword{datasets}
