% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Percent_of_Children_with_Elevated_Blood_Lead_Levels_DATA.R
\docType{data}
\name{Percent_of_Children_with_Elevated_Blood_Lead_Levels}
\alias{Percent_of_Children_with_Elevated_Blood_Lead_Levels}
\title{Percent of Children Tested (Aged 0-6) with Elevated Blood Lead Levels in Baltimore City}
\format{
## `Percent_of_Children_with_Elevated_Blood_Lead_Levels`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percent-of-children-tested-aged-0-6-with-elevated-blood-lead-levels-community-statistical-area/about>
}
\usage{
Percent_of_Children_with_Elevated_Blood_Lead_Levels
}
\description{
The number of children aged 0-6 that are found to either have elevated blood lead levels (10Mg/dL) or lead poisoning (20 Mg/dL) out of the number of children tested within an area in a calendar year.
}
\keyword{datasets}
