% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Percent_of_Employed_Population_with_Travel_Time_to_Work_of_45_Minutes_and_Over_DATA.R
\docType{data}
\name{Percent_of_Employed_Population_with_Travel_Time_to_Work_of_45_Minutes_and_Over}
\alias{Percent_of_Employed_Population_with_Travel_Time_to_Work_of_45_Minutes_and_Over}
\title{Percent of Employed Population with Travel Time to Work Over 45 Minutes in Baltimore City}
\format{
## `Percent_of_Employed_Population_with_Travel_Time_to_Work_of_45_Minutes_and_Over`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percent-of-employed-population-with-travel-time-to-work-over-45-minutes-community-statistical-area/about>
}
\usage{
Percent_of_Employed_Population_with_Travel_Time_to_Work_of_45_Minutes_and_Over
}
\description{
The percentage of commuters that spend more than 45 minutes travelling to work out of all commuters aged 16 and above.
}
\keyword{datasets}
