% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Percent_of_Households_with_No_Vehicles_Available_DATA.R
\docType{data}
\name{Percent_of_Households_with_No_Vehicles_Available}
\alias{Percent_of_Households_with_No_Vehicles_Available}
\title{Percent of Households with No Vehicle Available in Baltimore City}
\format{
## `Percent_of_Households_with_No_Vehicles_Available`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percent-of-households-with-no-vehicle-available-community-statistical-area/about>
}
\usage{
Percent_of_Households_with_No_Vehicles_Available
}
\description{
The percentage of households that do not have a personal vehicle available for use out of all households in an area.
}
\keyword{datasets}
