% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Percent_of_Population_that_Walks_to_Work_DATA.R
\docType{data}
\name{Percent_of_Population_that_Walks_to_Work}
\alias{Percent_of_Population_that_Walks_to_Work}
\title{Percent of Population that Walks to Work in Baltimore City}
\format{
## `Percent_of_Population_that_Walks_to_Work`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percent-of-population-that-walks-to-work-community-statistical-area/about>
}
\usage{
Percent_of_Population_that_Walks_to_Work
}
\description{
The percentage of commuters that walk to work out of all commuters aged 16 and above.
}
\keyword{datasets}
