% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Percentage_of_Students_Receiving_Free_or_Reduced_Meals_DATA.R
\docType{data}
\name{Percentage_of_Students_Receiving_Free_or_Reduced_Meals}
\alias{Percentage_of_Students_Receiving_Free_or_Reduced_Meals}
\title{Percentage of Students Receiving Free or Reduced Meals in Baltimore City}
\format{
## `Percentage_of_Students_Receiving_Free_or_Reduced_Meals`
A data frame with 55 obs. of  5 variables:
\describe{
  \item{CSA2010 }{Area Name in 2010 (Character String)}
  \item{susp10}{The percentage of students of any grade that are eligible for and received free or reduced school meals in 2010 (Numeric)}
  \item{susp11}{The percentage of students of any grade that are eligible for and received free or reduced school meals in 2011 (Numeric)}
  \item{susp12}{The percentage of students of any grade that are eligible for and received free or reduced school meals in 2012 (Numeric)}
  \item{susp13}{The percentage of students of any grade that are eligible for and received free or reduced school meals in 2013 (Numeric)}
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percentage-of-students-receiving-free-or-reduced-meals/about>
}
\usage{
Percentage_of_Students_Receiving_Free_or_Reduced_Meals
}
\description{
The percentage of students of any grade that are eligible for and received free or reduced school meals out of all public school students.
Eligibility for this program is based on the student's household income.
}
\keyword{datasets}
