% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Racial_Diversity_Index_DATA.R
\docType{data}
\name{Racial_Diversity_Index}
\alias{Racial_Diversity_Index}
\title{Baltimore City Racial Diversity Index}
\format{
## `Racial_Diversity_Index`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://data.baltimorecity.gov/datasets/bniajfi::racial-diversity-index-community-statistical-area/about>
}
\usage{
Racial_Diversity_Index
}
\description{
The percent chance that two people picked at random within an area will be of a different race/ethnicity. This number does not reflect which race/ethnicity is predominant within an area. The higher the value, the more racially and ethnically diverse an area. Source: U.S. Bureau of the Census, American Community Survey Years Available: 2010, 2011-2015, 2012-2016, 2013-2017, 2014-2018, 2015-2019
}
\keyword{datasets}
