% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Surface_Water_Quality_Data_1995_through_September_2022_DATA.R
\docType{data}
\name{Surface_Water_Quality_Data_1995_through_September_2022}
\alias{Surface_Water_Quality_Data_1995_through_September_2022}
\title{Surface Water Quality Data 1995 through Present}
\format{
## `Surface_Water_Quality_Data_1995_through_September_2022`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://data.baltimorecity.gov/datasets/surface-water-quality-data-1995-through-present/about>
}
\usage{
Surface_Water_Quality_Data_1995_through_September_2022
}
\description{
This dataset represents the monitoring for the water quality of the streams and Harbor in the City of Baltimore. The Environmental Compliance and Laboratory Services Division of the Bureau of Water and Wastewater is responsible for monitoring the quality of the streams and Harbor in the City of Baltimore. The purpose of this program is to evaluate the quality of our surface waters for any long-term trends and determine any capital investment programs or operational programs to implement. The results may be affected by a variety of factors: precipitation, land use changes, and human behavior. For more information:http://cleanwaterbaltimore.org/.
}
\keyword{datasets}
