% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Teen_Birth_Rate_per_1000_Females_DATA.R
\docType{data}
\name{Teen_Birth_Rate_per_1000_Females}
\alias{Teen_Birth_Rate_per_1000_Females}
\title{Teen Birth Rate per 1,000 Females (aged 15-19) in Baltimore City}
\format{
## `Teen_Birth_Rate_per_1000_Females`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::teen-birth-rate-per-1000-females-aged-15-19-community-statistical-area/about>
}
\usage{
Teen_Birth_Rate_per_1000_Females
}
\description{
The rate of female teens aged 15-19 that gave birth per 1,000 females aged 15-19.
}
\keyword{datasets}
