% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Baltimore_City_Employee_Salaries_DATA.R
\docType{data}
\name{Baltimore_City_Employee_Salaries}
\alias{Baltimore_City_Employee_Salaries}
\title{Baltimore City Employee Salaries}
\format{
`Baltimore_City_Employee_Salaries`
A data frame with 174067 obs. of  9 variables:
\describe{
  \item{OBJECTID}{Individual Row Database Identifier (Integer)}
  \item{Name}{Name of Individual (Character string)}
  \item{JobTitle}{Job Title of Individual (Character string)}
  \item{AgencyID}{Employer Database Identifier (Character string)}
  \item{AgencyName}{Name of Employer Agency (Character string)}
  \item{HireDate}{Date when Employee was hired (Date)}
  \item{AnnualSalary}{Annual/Yearly Employee Salary (Integer)}
  \item{GrossPay}{Gross Employee Pay (numeric)}
  \item{FiscalYear}{Fiscal Year of Obersvation Recording (Character string)}

}
}
\source{
<https://data.baltimorecity.gov/datasets/baltimore::baltimore-city-employee-salaries-2/about>
}
\usage{
Baltimore_City_Employee_Salaries
}
\description{
This dataset includes Baltimore City employee salaries and gross pay from fiscal year 2011 through last fiscal year and includes employees who were employed on June 30 of the last fiscal year. For fiscal years 2020 and prior, data are extracted from the ADP payroll system. For fiscal year 2022, the data are combined from the ADP system and the Workday enterprise resource planning system which now includes payroll.
}
\keyword{datasets}
