% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Customer_Service_Requests_2018_DATA.R
\docType{data}
\name{Customer_Service_Requests_2018}
\alias{Customer_Service_Requests_2018}
\title{Baltimore City 311 Customer Service Requests 2018}
\format{
## `Customer_Service_Requests_2018`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://data.baltimorecity.gov/datasets/baltimore::311-customer-service-requests-2018/about>
}
\usage{
Customer_Service_Requests_2018
}
\description{
This dataset represents the 311 call for service requests for the 2018 calendar year which includes service request types and status.
}
\keyword{datasets}
