% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/High_School_Completion_Rate_DATA.R
\docType{data}
\name{High_School_Completion_Rate}
\alias{High_School_Completion_Rate}
\title{High School Completion Rate in Baltimore City}
\format{
## `High_School_Completion_Rate`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::high-school-competion-rate-community-statistical-area/about>
}
\usage{
High_School_Completion_Rate
}
\description{
The percentage of 12th graders in a school year that successfully completed high school out of all 12th graders within an area.
Completers are identified as completing their program of study at the high school level and satisfying the graduation requirements for a Maryland High School Diploma or the requirements for a Maryland Certificate of Program Completion.
}
\keyword{datasets}
