% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Life_Expectancy_DATA.R
\docType{data}
\name{Life_Expectancy}
\alias{Life_Expectancy}
\title{Life Expectancy in Baltimore City}
\format{
## `Life_Expectancy`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::life-expectancy-community-statistical-area/about>
}
\usage{
Life_Expectancy
}
\description{
The average number of years a newborn can expect to live, assuming he or she experiences the currently prevailing rates of death through their lifespan.
}
\keyword{datasets}
