% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Percent_of_Population_that_Carpool_to_Work_DATA.R
\docType{data}
\name{Percent_of_Population_that_Carpool_to_Work}
\alias{Percent_of_Population_that_Carpool_to_Work}
\title{Percent of Population that Carpool to Work in Baltimore City}
\format{
## `Percent_of_Population_that_Carpool_to_Work`
A data frame with 55 obs. of  10 variables:
\describe{
  \item{CSA2010 }{Area Name (Character String)}
  \item{carpool11}{The percentage of commuters that carpool out of all commuters in 2011 (Numeric)}
  \item{carpool12}{The percentage of commuters that carpool out of all commuters in 2012 (Numeric)}
  \item{carpool13}{The percentage of commuters that carpool out of all commuters in 2013 (Numeric)}
  \item{carpool14}{The percentage of commuters that carpool out of all commuters in 2014 (Numeric)}
  \item{carpool15}{The percentage of commuters that carpool out of all commuters in 2015 (Numeric)}
  \item{carpool16}{The percentage of commuters that carpool out of all commuters in 2016 (Numeric)}
  \item{carpool17}{The percentage of commuters that carpool out of all commuters in 2017 (Numeric)}
  \item{carpool18}{The percentage of commuters that carpool out of all commuters in 2018 (Numeric)}
  \item{carpool19}{The percentage of commuters that carpool out of all commuters in 2019 (Numeric)}
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percent-of-population-that-carpool-to-work-community-statistical-area/about>
}
\usage{
Percent_of_Population_that_Carpool_to_Work
}
\description{
The percentage of commuters that carpool out of all commuters aged 16 and above.
}
\keyword{datasets}
