% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Percent_of_Population_that_Walks_to_Work_DATA.R
\docType{data}
\name{Percent_of_Population_that_Walks_to_Work}
\alias{Percent_of_Population_that_Walks_to_Work}
\title{Percent of Population that Walks to Work in Baltimore City}
\format{
## `Percent_of_Population_that_Walks_to_Work`
A data frame with 55 obs. of  10 variables:
\describe{
  \item{CSA2010 }{Area Name (Character String)}
  \item{walked11}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2011 (Numeric)}
  \item{walked12}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2012 (Numeric)}
  \item{walked13}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2013 (Numeric)}
  \item{walked14}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2014 (Numeric)}
  \item{walked15}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2015 (Numeric)}
  \item{walked16}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2016 (Numeric)}
  \item{walked17}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2017 (Numeric)}
  \item{walked18}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2018 (Numeric)}
  \item{walked19}{The percentage of commuters that walk to work out of all commuters aged 16 and above in 2019 (Numeric)}
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percent-of-population-that-walks-to-work-community-statistical-area/about>
}
\usage{
Percent_of_Population_that_Walks_to_Work
}
\description{
The percentage of commuters that walk to work out of all commuters aged 16 and above.
}
\keyword{datasets}
