% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Percentage_of_Students_Receiving_Free_or_Reduced_Meals_DATA.R
\docType{data}
\name{Percentage_of_Students_Receiving_Free_or_Reduced_Meals}
\alias{Percentage_of_Students_Receiving_Free_or_Reduced_Meals}
\title{Percentage of Students Receiving Free or Reduced Meals in Baltimore City}
\format{
## `Percentage_of_Students_Receiving_Free_or_Reduced_Meals`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percentage-of-students-receiving-free-or-reduced-meals/about>
}
\usage{
Percentage_of_Students_Receiving_Free_or_Reduced_Meals
}
\description{
The percentage of students of any grade that are eligible for and received free or reduced school meals out of all public school students.
Eligibility for this program is based on the student's household income.
}
\keyword{datasets}
