% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Percentage_of_Students_Suspended_or_Expelled_During_School_Year_DATA.R
\docType{data}
\name{Percentage_of_Students_Suspended_or_Expelled_During_School_Year}
\alias{Percentage_of_Students_Suspended_or_Expelled_During_School_Year}
\title{Percentage of Students Suspended or Expelled During School Year in Baltimore City}
\format{
## `Percentage_of_Students_Suspended_or_Expelled_During_School_Year`
A data frame with BLA rows and BLA columns:
\describe{
  \item{BLA}{BLA}
  \item{BLA, BLA}{BLABLABLA}
  \item{BLA}{BLA}
  ...
}
}
\source{
<https://vital-signs-bniajfi.hub.arcgis.com/datasets/bniajfi::percentage-of-students-suspended-or-expelled-during-school-year/about>
}
\usage{
Percentage_of_Students_Suspended_or_Expelled_During_School_Year
}
\description{
The percentage of students of any grade level that were formally suspended or expelled for any reason during the school year out of all public school students within an area.
}
\keyword{datasets}
